COMPLETED APP

+++ src/main/java/com/goldman/timetracker/repository/DummyDataRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.goldman.timetracker.entity.DummyData;
 
-public interface DummyDataRepository extends CrudRepository<DummyData, Integer>  {
+public interface DummyDataRepository extends CrudRepository<DummyData, Long>  {
 	DummyData findByDescription(String description);
 }
